Added new New Entry Box to Settings Editor
When editing the settings to a question in studio
the editor (Person editing) now
has the option of specifying that
the show answer button be shown to students only
after they have attempted answering the question
at least 'X' times, where x is a value that the
editor specifies. To this end, the show answer
dropdown has a new option, 'After X Attempts',
whose name has been changed from 'Attempted x Times'.
There is also a new entry-box called Value of X
in which the editor specifies the value of x.