Remove Old Sneakpeek Implementation
Gets rid of the previous impl. of the
sneak peek feature and all references to
any of its aspects in the codebase.
Changes in this commit also fix almost
all (We still have 3 Errors and 3 Test Fails) Errors and test
fails resulting from the purge.

Introduce djangolazysignup and sneakpeek

Fixes some of the remaining errors and introduces
lazy-signup as a replacement for the previous
impl. of sneakpeek. Creates a new app (called sneakpeek)
that will house code related to the new impl.

Create Blocking Middleware.

Add implementation for middleware to prevent
sneakpeek users from accessing studio. Introduces a
new button on the course about page that allows
users to access course content as sneak peak users.
Clicking the button leads to the creation of a new
non registered user.